add exception in CommentService

+++ back/src/main/java/com/openclassrooms/back/services/CommentService.java
@@ -1,6 +1,7 @@
 package com.openclassrooms.back.services;
 
 import com.openclassrooms.back.dto.CommentRequest;
+import com.openclassrooms.back.exceptions.NotFoundException;
 import com.openclassrooms.back.models.Comment;
 import com.openclassrooms.back.repositories.CommentRepository;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +27,9 @@ public class CommentService {
      * @return la liste des commentaires
      */
     public List<Comment> getCommentsByPostId(Long postId) {
+        if (!postService.existsById(postId)) {
+            throw new NotFoundException("Comments with Post id " + postId + " not found");
+        }
         return commentRepository.findByPostId(postId);
     }
 
@@ -35,7 +39,8 @@ public class CommentService {
      * @return le commentaire
      */
     public Comment getCommentById(Long id) {
-        return commentRepository.findById(id).get();
+        return commentRepository.findById(id)
+                .orElseThrow(() -> new NotFoundException("Comment with id " + id + " not found"));
     }
 
     /**
@@ -44,6 +49,18 @@ public class CommentService {
      * @return le commentaire créé
      */
     public Comment createComment(CommentRequest commentRequest) {
+        if (commentRequest.getContent() == null || commentRequest.getContent().isEmpty()) {
+            throw new NotFoundException("Content must not be null or empty");
+        }
+
+        if (!userService.existsById(commentRequest.getUserId())) {
+            throw new NotFoundException("User with id " + commentRequest.getUserId() + " not found");
+        }
+
+        if (!postService.existsById(commentRequest.getPostId())) {
+            throw new NotFoundException("Post with id " + commentRequest.getPostId() + " not found");
+        }
+
         Comment comment = new Comment();
         comment.setContent(commentRequest.getContent());
         comment.setUser(userService.getUserById(commentRequest.getUserId()));
@@ -56,6 +73,9 @@ public class CommentService {
      * @param id l'id du commentaire
      */
     public void deleteComment(Long id) {
+        if (!commentRepository.existsById(id)) {
+            throw new NotFoundException("Comment with id " + id + " not found");
+        }
         commentRepository.deleteById(id);
     }
 }
